requirements.txt
opencv-python
smtplib
email
ultralytics
streamlit
python-dotenv
